mod_ehr/requirements.txt
asttokens==2.4.1
attrs==23.2.0
aws-cdk-lib
aws-cdk.asset-awscli-v1
aws-cdk.asset-kubectl-v20
aws-cdk.asset-node-proxy-agent-v6
boto3==1.34.52
botocore==1.34.52
cattrs==23.2.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
constructs
cryptography==42.0.5
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
idna==3.6
importlib-resources==6.1.1
ipython==8.21.0
jedi==0.19.1
jmespath==1.0.1
jsii==1.96.0
jwt==1.3.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
publication==0.0.3
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
stack-data==0.6.3
traitlets==5.14.1
typeguard==2.13.3
typing_extensions==4.9.0
urllib3==2.0.7
wcwidth==0.2.13
xmltodict==0.13.0
